+++ src/sll/api/math_api.c
@@ -116,7 +116,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_factor_t* sll_math_factors(sll_size_t n,sl
 		} while (!(n%5));
 		PUSH_FACTOR(5,i);
 	}
-	wide_integer_t f=7;
+	__SLL_U64 f=7;
 	if (n>48){
 		while (1){
 			WHEEL_STEP(4);
@@ -212,11 +212,11 @@ __API_FUNC(math_combinations){
 	if (b==1){
 		return 1;
 	}
-	wide_integer_t av=(wide_integer_t)a;
-	wide_integer_t bv=(wide_integer_t)b;
-	wide_integer_t o=1;
-	wide_integer_t i=a-b;
-	wide_integer_t j=2;
+	__SLL_U64 av=(__SLL_U64)a;
+	__SLL_U64 bv=(__SLL_U64)b;
+	__SLL_U64 o=1;
+	__SLL_U64 i=a-b;
+	__SLL_U64 j=2;
 	do{
 		i++;
 		o*=i;
@@ -277,7 +277,7 @@ __API_FUNC(math_factorial){
 	if (a<0){
 		SLL_UNIMPLEMENTED();
 	}
-	wide_integer_t o=1;
+	__SLL_U64 o=1;
 	while (a>1){
 		o*=a;
 		a--;
@@ -345,15 +345,15 @@ __API_FUNC(math_int_pow){
 	if (!b){
 		return 1;
 	}
-	wide_integer_t o=1;
+	__SLL_U64 o=1;
 	if (a<0){
 		a=-a;
 		if (b&1){
 			o=-1;
 		}
 	}
-	wide_integer_t v=(wide_integer_t)a;
-	wide_integer_t e=(wide_integer_t)b;
+	__SLL_U64 v=(__SLL_U64)a;
+	__SLL_U64 e=(__SLL_U64)b;
 	while (1){
 		if (e&1){
 			o*=v;
@@ -373,12 +373,12 @@ __API_FUNC(math_int_sqrt){
 	if (a<0){
 		SLL_UNIMPLEMENTED();
 	}
-	wide_integer_t v=(wide_integer_t)a;
-	wide_integer_t i=v>>1;
+	__SLL_U64 v=(__SLL_U64)a;
+	__SLL_U64 i=v>>1;
 	if (!i){
 		return v;
 	}
-	wide_integer_t j=i;
+	__SLL_U64 j=i;
 	do{
 		i=j;
 		j=(i+v/i)>>1;
@@ -416,9 +416,9 @@ __API_FUNC(math_permutations){
 	if (b==1){
 		return a;
 	}
-	wide_integer_t v=(wide_integer_t)a;
-	wide_integer_t o=1;
-	wide_integer_t i=a-b;
+	__SLL_U64 v=(__SLL_U64)a;
+	__SLL_U64 o=1;
+	__SLL_U64 i=a-b;
 	do{
 		i++;
 		o*=i;