+++ src/sll/include/sll/_size_types.h
@@ -1,0 +1,30 @@
+#ifndef __SLL__SIZE_TYPES_H__
+#define __SLL__SIZE_TYPES_H__ 1
+
+
+
+#define __SLL_S8 signed char
+#define __SLL_U8 unsigned char
+#define __SLL_S16 signed short int
+#define __SLL_U16 unsigned short int
+#define __SLL_S32 signed int
+#define __SLL_U32 unsigned int
+#define __SLL_S64 signed long long int
+#define __SLL_U64 unsigned long long int
+
+
+
+#define __SLL_F32 float
+#define __SLL_F64 double
+
+
+
+
+#define __SLL_U8_MAX 255
+#define __SLL_U16_MAX 65535
+#define __SLL_U32_MAX 4294967295u
+#define __SLL_U64_MAX 18446744073709551615ull
+
+
+
+#endif