+++ src/sll/include/sll/identifier.h
@@ -11,7 +11,6 @@
 
 
 
-#define SLL_MAX_SHORT_IDENTIFIER_LENGTH 15
 #define SLL_IDENTIFIER_GET_ARRAY_ID(i) ((i)&0xf)
 #define SLL_IDENTIFIER_GET_ARRAY_INDEX(i) ((i)>>4)
 #define SLL_IDENTIFIER_ADD_INDEX(i,j) ((i)+((j)<<4))