+++ src/sll/include/sll/string.h
@@ -39,12 +39,12 @@
 #define SLL_STRING_FORMAT_PADDING(v,l) \
     do{ \
         if (((l)&15)<8){ \
-            (*((uint64_t*)((v)+((l)&0xfffffffffffffff0ull))))&=(1ull<<(((l)&15)<<3))-1; \
+            (*((__SLL_U64*)((v)+((l)&0xfffffffffffffff0ull))))&=(1ull<<(((l)&15)<<3))-1; \
         } \
-        (*((uint64_t*)((v)+((l)&0xfffffffffffffff0ull)+8)))&=(1ull<<((((l)&15)-8)<<3))-1; \
+        (*((__SLL_U64*)((v)+((l)&0xfffffffffffffff0ull)+8)))&=(1ull<<((((l)&15)-8)<<3))-1; \
     } while (0)
 #define SLL_STRING_HEX_ESCAPE(c) ((c)<32||(c)>126)
-#define SLL_STRING_INSERT_POINTER_STATIC(nm,i,s) sll_string_insert_pointer_length(SLL_CHAR((nm)),sizeof(nm)/sizeof(char)-1,(i),(s))
+#define SLL_STRING_INSERT_POINTER_STATIC(nm,i,s) sll_string_insert_pointer_length(SLL_CHAR(nm),sizeof(nm)/sizeof(char)-1,(i),(s))