+++ src/sll/parse.c
@@ -3,6 +3,7 @@
 #include <sll/common.h>
 #include <sll/data.h>
 #include <sll/file.h>
+#include <sll/identifier.h>
 #include <sll/ift.h>
 #include <sll/memory.h>
 #include <sll/node.h>