+++ src/sll/print.c
@@ -2,6 +2,7 @@
 #include <sll/assembly.h>
 #include <sll/common.h>
 #include <sll/file.h>
+#include <sll/identifier.h>
 #include <sll/memory.h>
 #include <sll/node.h>
 #include <sll/object.h>